+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.3
+- Fixed an error in the --create-web-profile command
+- The web profiles table now shows the number of stored webmaps
+- Added a command to delete webmaps
+- Added a command to delete web profiles
+
+
 ## 0.2
 - Fixed a typo in the --help screen
 - Webmap comparision function added